삭제시 문자 보내기 기능 추가 -> locationList_Delete Ajax 확인 요

+++ src/main/java/co/soft/controller/FileController.java
@@ -39,7 +39,9 @@ public class FileController {
 	public String deleteFile(Model model, HttpServletRequest request) {
 		Long boardidx = Long.parseLong(request.getParameter("boardidx"));
 		FileInfoBean fileidx= fileService.getFile(boardidx);
-		fileService.deleteFile(fileidx);
+		if(fileidx != null) {
+			fileService.deleteFile(fileidx);
+		}
 		return "/location/Location_Delete_Success";
 	}
